Address PR 43 review round 2: off-main-actor I/O, slider undo, cleanup
- importLookPreset: resolve the sidecar LUT bookmark inside the same
  detached task that reads/decodes the preset, so no disk I/O runs on the
  main actor. applyLookPreset now takes a pre-resolved lutBookmark.
- resolvePresetLUT: make nonisolated static so it is callable from the
  detached task without hopping back to the main actor.
- applyBuiltInLookPreset: pass lutBookmark: nil.
- exportLookPreset: push the disk write off the main actor via a detached
  task; encode on the caller, write in the background.
- InspectorView: commit the coalesced undo group on drag-end for all
  Grain/Halation/Vignette sliders (onEditingChanged), matching the rest of
  the inspector.
- EffectCompositor.applyHalation: drop the redundant CGFloat round-trip on
  the Float threshold.

+++ LocalCut Studio/EffectCompositor.swift
@@ -646,11 +646,11 @@ final class EffectCompositor: NSObject, AVVideoCompositing {
         let strength = params.strength(at: time)
         guard strength > 0, params.radius > 0 else { return image }
 
-        let threshold = CGFloat(params.threshold)
+        let threshold = params.threshold
         let brightPass = CIFilter.colorControls()
         brightPass.inputImage = image
-        brightPass.brightness = -Float(threshold * 0.45)
-        brightPass.contrast = Float(2.5 + threshold * 4)
+        brightPass.brightness = -threshold * 0.45
+        brightPass.contrast = 2.5 + threshold * 4
         brightPass.saturation = 0.85
 
         let blur = CIFilter.gaussianBlur()
